requirements.txt
aiosqlite==0.21.0
alembic==1.17.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
Authlib==1.6.5
beautifulsoup4==4.14.2
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
cloudpickle==3.1.2
crashtest==0.4.1
cryptography==46.0.3
distlib==0.3.9
distro==1.9.0
docstring_parser==0.17.0
dulwich==0.22.8
fastapi==0.123.10
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.21.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.128.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.6.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.55.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.22.0
more-itertools==10.7.0
msgpack==1.1.0
narwhals==2.13.0
numpy==2.3.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.3
pbs-installer==2025.5.17
pexpect==4.9.0
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
poetry-plugin-shell==1.0.1
proto-plus==1.26.1
protobuf==6.33.2
ptyprocess==0.7.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlparse==0.5.3
sse-starlette==3.0.3
starlette==0.50.0
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.3
trove-classifiers==2025.5.9.12
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
virtualenv==20.31.2
watchdog==6.0.0
websockets==15.0.1
xattr==1.1.4
zipp==3.23.0
zstandard==0.23.0
